requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==19.3.0
    # via
    #   automat
    #   jsonschema
    #   service-identity
    #   treq
    #   twisted
automat==0.8.0
    # via twisted
certifi==2019.11.28
    # via
    #   requests
    #   sentry-sdk
cffi==1.13.2
    # via cryptography
chardet==3.0.4
    # via requests
constantly==15.1.0
    # via twisted
contextlib2==0.6.0.post1
    # via schema
cryptography==3.2.1
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.1.0
    # via
    #   parsel
    #   scrapy
defusedxml==0.6.0
    # via odfpy
et-xmlfile==1.0.1
    # via openpyxl
flattentool==0.14.0
    # via -r requirements.in
hyperlink==19.0.0
    # via
    #   treq
    #   twisted
idna==2.8
    # via
    #   hyperlink
    #   requests
    #   twisted
ijson==3.1.1
    # via -r requirements.in
importlib-metadata==1.6.1
    # via jsonschema
incremental==17.5.0
    # via
    #   treq
    #   twisted
itemadapter==0.1.1
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.0.3
    # via scrapy
jdcal==1.4.1
    # via openpyxl
jmespath==0.10.0
    # via itemloaders
jsonpointer==2.0
    # via -r requirements.in
jsonref==0.2
    # via
    #   -r requirements.in
    #   flattentool
    #   ocdsmerge
jsonschema==3.2.0
    # via -r requirements.in
lxml==4.6.2
    # via
    #   flattentool
    #   parsel
    #   scrapy
ocdsmerge==0.6.4
    # via -r requirements.in
odfpy==1.4.1
    # via flattentool
openpyxl==3.0.5
    # via flattentool
parsel==1.5.2
    # via
    #   itemloaders
    #   scrapy
protego==0.1.16
    # via scrapy
pyasn1-modules==0.2.7
    # via service-identity
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pycparser==2.19
    # via cffi
pydispatcher==2.0.5
    # via scrapy
pyhamcrest==1.9.0
    # via twisted
pyopenssl==19.1.0
    # via
    #   scrapy
    #   twisted
pyrsistent==0.16.0
    # via jsonschema
pytz==2020.1
    # via flattentool
queuelib==1.5.0
    # via scrapy
rarfile==3.1
    # via -r requirements.in
requests==2.22.0
    # via
    #   ocdsmerge
    #   treq
rfc3987==1.3.8
    # via -r requirements.in
schema==0.7.2
    # via flattentool
scrapy==2.3.0
    # via
    #   -r requirements.in
    #   scrapyd
    #   scrapyd-client
scrapyd-client==1.1.0
    # via -r requirements.in
scrapyd==1.2.1
    # via -r requirements.in
sentry-sdk==0.14.4
    # via -r requirements.in
service-identity==18.1.0
    # via
    #   scrapy
    #   twisted
six==1.13.0
    # via
    #   automat
    #   cryptography
    #   jsonschema
    #   parsel
    #   protego
    #   pyhamcrest
    #   pyopenssl
    #   pyrsistent
    #   scrapyd
    #   scrapyd-client
    #   treq
    #   w3lib
treq==20.9.0
    # via -r requirements.in
twisted[tls]==20.3.0
    # via
    #   -r requirements.in
    #   scrapy
    #   scrapyd
    #   treq
urllib3==1.25.7
    # via
    #   requests
    #   sentry-sdk
w3lib==1.21.0
    # via
    #   itemloaders
    #   parsel
    #   scrapy
xmltodict==0.12.0
    # via flattentool
zipp==3.1.0
    # via importlib-metadata
zope.interface==5.2.0
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools

# Bidhive Django libraries
Django==3.1
celery==5.0.5
django-cors-headers==3.4.0
django-extensions==3.0.5
django-filter==2.3.0
django-macros==0.4.0
djangorestframework==3.11.1
django-autoslug==1.9.8
drf-nested-routers==0.91
pyquery==1.4.1
psycopg2-binary==2.8.6
redis==3.5.3